# swagger_docs.yaml
openapi: 3.0.2
info:
  title: StoryBoom AI API
  description: |
    API for StoryBoom AI - Case Study Generation Platform
    
    This API provides endpoints for:
    - User authentication and management
    - Case study creation and management
    - Interview management (provider and client)
    - Media generation (videos, podcasts)
    - Feedback collection and analysis
    - Metadata analysis
    
    ## Authentication
    Most endpoints require authentication. Use session-based authentication by logging in first.
    
    ## Rate Limits
    API calls are limited to prevent abuse. Contact support if you need higher limits.
    
    ## IP Access
    For production access, contact support to whitelist your IP address.
  version: 1.0.0
  contact:
    name: StoryBoom AI Support
    email: storyboomai@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:10000
    description: Development server
  - url: https://your-production-domain.com
    description: Production server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Session-based authentication (login required)
...
paths:
...
  /api/feedback/submit:
    post:
      tags:
        - Feedback
      summary: Submit feedback
      description: Submit user feedback with optional rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: Feedback content
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Optional rating (1-5)
                feedback_type:
                  type: string
                  default: general
                  description: Type of feedback
      responses:
        '200':
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  feedback_id:
                    type: integer
                  feedback_summary:
                    type: string
...